Bug: togglePause on re-opening popup fixed
The state did not load correctly after closing the popup and re-opening it again. The state would instead be the default state (isPaused=false) because the state was saved to _isPaused, which is wrong! After removing the lowdash and saving the state back to the actual state isPaused the pause-button is working again correctly.